requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
langchain==0.3.28
langchain-community==0.3.27
chromadb
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
pandas==2.2.2
requests==2.32.2
tiktoken==0.7.0
google-genai
langchain-google-genai
psycopg2-binary==2.9.9